refactor async cleaning logic out into core so caches transparently take advantage of it (and clean code no longer needs to know about it)
